restructure of MetaDataView to reduce number of re-renders
